code/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-pubsub
google-cloud-secret-manager==2.0.0
sqlalchemy
pymysql